Rename trajectory_parser to trajectory_loader
