+++ analyzer/__init__.py
@@ -2,7 +2,7 @@ from . import _version
 from .data.geometry import Geometry
 from .data.trajectory_data import TrajectoryData, TrajectoryUnit
 from .io.geometry_parser import parse_geometry
-from .io.trajectory_parser import load_trajectory
+from .io.trajectory_loader import load_trajectory
 from .methods.density_calculator import (
     compute_classic_density,
     compute_passing_density,
