+++ analyzer/io/trajectory_loader.py
