+++ tests/unit_tests/io/test_trajectory_loader.py
@@ -6,8 +6,8 @@ import pytest
 from numpy import dtype
 
 from analyzer.data.trajectory_data import TrajectoryUnit
-from analyzer.io.trajectory_parser import *
-from analyzer.io.trajectory_parser import (
+from analyzer.io.trajectory_loader import *
+from analyzer.io.trajectory_loader import (
     _load_trajectory_data,
     _load_trajectory_meta_data,
 )
